presentation-api/controlling-ua/getAvailability_success-manual.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Getting the presentation displays availability information.</title>
<link rel="author" title="Marius Wessel" href="http://www.fokus.fraunhofer.de">
<link rel="author" title="Tomoyuki Shimizu" href="https://github.com/tomoyukilabs">
<link rel="help" href="http://w3c.github.io/presentation-api/#dfn-presentation-display-availability">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="common.js"></script>
<style>
#guide {
    display: none;
}
</style>

<div id="notice">
    <p>Please click the button below to start the manual test,<br>
    when any presentation display becomes available.</p>
    <button id="presentBtn">Start Presentation Test</button>
    <div id="guide">
        <p>If your browser is unable to monitor the presentation displays availability<br>
        (e.g., because the user has disabled this feature), please click the left button below;<br>
        otherwise, please click the right button below.</p>
        <button id="skipBtn">Monitoring is disabled</button>
        <button id="continueBtn">Continue the test</button>
    </div>
</div>
<p>The test passes if a "PASS" result appears.</p>

<script>
    // disable the timeout function for the tests
    setup({explicit_timeout: true});

    // ------------
    // DOM Elements
    // ------------
    var notice      = document.getElementById('notice'),
        presentBtn  = document.getElementById('presentBtn'),
        guide       = document.getElementById('guide'),
        skipBtn     = document.getElementById('skipBtn'),
        continueBtn = document.getElementById('continueBtn');

    // ---------------------------------
    // Screen Availability Tests - begin
    // ---------------------------------
    presentBtn.onclick = function() {
        presentBtn.disabled = true;

        promise_test(function(t) {
        // clean up the instruction notice when the test ends
            t.add_cleanup(function() {
                notice.parentNode.removeChild(notice);
            });

            // instance of PresentationRequest
            var request = new PresentationRequest(presentationUrls);
            assert_true(request instanceof PresentationRequest, 'The request is an instance of PresentationRequest.');

            let availability1;

            return request.getAvailability()
                    .then(function (availability) {
                        // instance of PresentationAvailability
                        assert_true(availability instanceof PresentationAvailability, 'The promise is resolved with an instance of PresentationAvailability.');
                        // PresentationAvailability.value is set as boolean
                        assert_true(typeof availability.value == 'boolean', 'The availability has an boolean value.');

                        availability1 = availability;

                        // wait for clicking either buttons
                        if(!availability1.value) {
                            guide.style.display = 'block';
                            var skipWatcher     = new EventWatcher(t, skipBtn,     'click'),
                                continueWatcher = new EventWatcher(t, continueBtn, 'click');
                            return Promise.race([
                                skipWatcher    .wait_for('click').then(function() { return true; }),
                                continueWatcher.wait_for('click').then(function() {
                                    assert_true(availability1.value, 'The availability value is true when any presentation display is available.');
                                    return false;
                                })
                            ]);
                        }
                        // if any presentation display is already available, proceed to the next step
                        else
                            return false;
                    }, function(err) {
                        // the Promise is rejected with NotSupportedError if the browser can find displays only when starting a connection
                        assert_equals(err.name, 'NotSupportedError', 'getAvailability() rejects a Promise with a NotSupportedError exception, if the user agent can find presentation displays only when starting a connection.')
                    }).then(function(disabled) {
                        skipBtn    .disabled = true;
                        continueBtn.disabled = true;

                        // If monitoring feature is unable (e.g., disabled by a user), abort all the remaining steps
                        if(!disabled) {
                            return request.getAvailability()
                                    .then(function (availability2) {
                                        // the same PresentationAvailability object is obtained for the same presentation request
                                        assert_true(availability1 === availability2, 'getAvailability() resolves a Promise with the same PresentationAvailability object for the same presentation request.');
                                    });
                        }
                    });
        });
    };

</script>
